Reduce the number of state rpc calls in audit
This commit reduces the number of state RPC calls made during audit
from 7 for each subcloud to 1 by performing a single batch request
at the end of a subcloud's audit process in the worker, updating
availability and/or the endpoint status at once.

Test plan:
All of the tests were performed both in small and large DC systems.
1. PASS: Manage a subcloud and verify that its sync status is
updated to in-sync.
2. PASS: Unmanage a subcloud and verify that its sync status
is updated to unknown.
3. PASS: Turn off a subcloud and verify that its availability
is updated to offline.
4. PASS: Turn on a subcloud and verify that its availability
is updated to online.
5. PASS: Validate the state.log and ensure that only one RPC
call is made for each subcloud.

Story: 2011106
Task: 50374

Change-Id: I9d060444315deb4d6e5e735c9854822f6faf522e
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>