Improve subcloud endpoint update
This commit removes the dcorch endpoint update call during network
reconfiguration. After the dcorch integration with the
OptimizedOpenStackDriver [1], the subcloud endpoints are not cached
anymore and instead are built using the subcloud management IP. So the
OptimizedOpenStackDriver is only used to get the endpoints of the
master region (RegionOne), which doesn't change during network
reconfiguration.

This commit also fixes an issue where the dcmanager-manager service's
own endpoint cache was not updated during network reconfiguration.

It also fixes or change the OptimizedOpenStack driver usage in a few
places that were not modified in previous commits.

Test Plan:
1. PASS - To test platform sync, run dns-modify with and without
          --os-region-name SystemController and verify that when using
          the option, dcorch-proxy trigger a sync request immediatelly
          and that without the option it still gets triggered by dcorch
          audit. Verify that the DNS change correctly propagated to the
          subcloud;
2. PASS - To test identity sync, repeat the previous test but instead
          of changing the DNS, create a new user using the 'openstack
          user create/delete' command;
3. PASS - To test the removal of the endpoint update, modify the DNS
          without '--os-region-name SystemController', run the network
          reconfig and verify that the dcorch sync uses the new IP and
          the DNS changes are correctly synced with the subcloud;
4. PASS - Run a fernet key rotation test and verify that it still works
          as expected.

[1]: https://review.opendev.org/c/starlingx/distcloud/+/919959

Story: 2011106
Task: 50417

Change-Id: Ic06c6f6cfb12b440e36f72eb29bbffaf8988b921
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>