Integrate dcagent with dcmanager-audit
This commit integrates the newly created dcagent with dcmanager-audit.
Currently, dcagent is disabled and will remain disabled until
installed and enabled in subclouds. To enable, the return call in the
function subcloud_has_dcagent in dccommon/utils needs to be removed,
as well as removing the comments that check if dcagent should be
called.

NOTE: patch and load audits will be deprecated in the next major
release, which is why they are not performed by dcagent. Until then,
they will remain enabled.

Test plan:
  - PASS: Run dcmanager audit with dcagent. Verify only one call is
          made to audit the subcloud and the response include the
          correct sync status.
  - PASS: Run dcmanager audit without dcagent. Verify the audit
          works as expected querying each individual endpoint.
  - PASS: Verify dcmanager can audit a subcloud with dcagent and
          another one with old audit workflow at the same time and
          verify both subclouds get to 'in-sync' status.
  - PASS: Induce a status where the sync status of the subclouds'
          endpoints change (changing cert-id of kube rootca).
          Verify that the audit and communication to dcmanager-state
          happen within the 30s audit interval.

Story: 2011106
Task: 50560

Change-Id: Ic330a18daefeb6f12fa7589ae6ba82cafd30f298
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>