Cache RegionOne k8s versions during DC orchestrator k8s upgrade
The RegionOne Kubernetes version was being requested two times per
subcloud during a kubernetes upgrade orchestration, one time during
the pre-check state and another during the create VIM state.
This can cause issues when running the orchestration with a large
number of parallel subclouds.

This commit moves the cache implementation [1] to the orchestrator
folder, and then use it for the Kubernetes orchestrator to cache the
RegionOne k8s versions.

Test Plan:
1. PASS - Run the Kubernetes upgrade strategy and verify that it
          completes successfully and that the RegionOne k8s versions
          sysinv query is called a single time during the
          strategy apply.

Story: 2011106
Task: 50491

[1]: https://review.opendev.org/c/starlingx/distcloud/+/841093

Change-Id: I56c353977ef5476942d9e6090b29fbbe4330c553
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>