Disable patch-strategy for current release and onwards
This commit introduces changes to disable the patch-strategy for the
current release and onwards. The patch-strategy is still active for n-1
subclouds to allow the application of a specific patch to enable the new
USM feature through the usage of the --patch argument.

To reflect the current behavior of the patch orchestration, the
FinishingPatchStrategyState and PatchJobData were removed, as they
are no longer required to apply the USM patch.

Test Plan:

PASS - Execute patch-strategy create/apply for n-1 subclouds
- dcmanager patch-strategy create --upload-only --patch test.patch
- dcmanager patch-strategy apply
- dcmanager patch-strategy delete
- dcmanager patch-strategy create --patch test.patch
- dcmanager patch-strategy apply

Failure Path:

PASS - Unable to execute `patch-strategy create` for current release
- dcmanager patch-strategy create --patch test.patch
  - ERROR (app) Unable to process the contained instructions Unable to
    create strategy of type 'patch': Bad strategy request: Subcloud
    subcloud1-stx-latest has the same software version than the system
    controller. The --patch option only works with n-1 subclouds
PASS - Unable to execute `patch-strategy create` without a patch file
- dcmanager patch-strategy create
  - ERROR (app) The server could not comply with the request since it
    is either malformed or otherwise incorrect. Patch paramater is
    required for strategy type: patch.

Story: 2010676
Task: 50609

Change-Id: Id7b57fd6b22c92d561c1faeedb017002a58a66b6
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>