Adapt to new Keystone session creation timeouts
Following the inclusion of the connect and read timeouts for the
Keystone client in [1], this commit updates its use in the distcloud
repository. This way, it becomes faster to check if a subcloud is
online or offline by setting a shorter value to the connection timeout,
while keeping the read timeout longer.

[1] https://review.opendev.org/c/starlingx/upstream/+/921973

Test Plan:

- PASS: Build distributedcloud package
- PASS: Build ISO image
- PASS: Install and bootstrap a system controller
- PASS: Install and add a subcloud
- PASS: Manage the subcloud and verify it is in-sync
- PASS: Manual check of periodic audit shows no errors

Story: 2011106
Task: 50365

Change-Id: Ie334006cd63c615117c0d0918b7a63047e3a5fca
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>