Validate the release parameter for prestage
A validation mechanism is added for the release format
and whether the release passed as a parameter is deployed
on the system controller.

If the release parameter is not provided, the current
release of the system controller will be used.

Only release format such as MM.mm is supported.

If the format for a patch (MM.mm.pp) is provided,
the process will return an error, indicating that the
format must be MM.mm.

If the release being provided does not exist within the
list of deployed releases, an error will be returned,
indicating that the specified release is not supported.

When there are more than one minor release (MM.mm.pp)
deployed on the system controller, the prestage process
will be based on the MM.mm version, copying all the
existing commits on the system controller to the subcloud
for that major release.

Additionally, the variable naming convention has been
updated to use "software_version" for release version,
leaving "software_major_release" ineffective.

The "prestage_install" variable is also renamed
"prestage_reason" to avoid confusion with values
(for_install, for_sw_deploy).

Test Plan:
PASS: - Deploy a 24.09 DC with at least 1 subcloud.
      - Upload 24.09.1 patch release on the system controller
      - Prestage the subcloud supplying release patch format:
          dcmanager subcloud prestage --release 24.09.1
          --for-sw-deploy subcloudX
      - Verify error message:
          "Specified release format is not supported. Version
          format must be MM.mm"

PASS: - Deploy a 24.09 DC with at least 1 subcloud.
      - Upload and deploy 24.09.1 patch release on the system
        controller
      - Prestage the subcloud supplying not deployed release.
          Assuming 24.09.0 and 24.09.1 are deployed:
          dcmanager subcloud prestage --release 24.12
          --for-sw-deploy subcloudX
      - Verify error message:
          "The requested software version was not installed in
          the system controller, cannot prestage for software
          deploy"

PASS: - Deploy a 24.09 DC with at least 1 subcloud.
      - Upload and deploy 24.09.1 patch release on the system
        controller
      - Prestage the subcloud using command:
          dcmanager subcloud prestage --release 24.09
          --for-sw-deploy subcloudX
      - Verify that 24.09.1 ostree commit and metadata is
        prestaged on the subcloud as the patch release has
        been deployed on the system controller
      - Check the software list command output to see the
        release states.

PASS: - Deploy a 24.09 DC with at least 1 subcloud.
      - Upload and deploy 24.09.1 patch release on the system
        controller
      - Prestage the subcloud using command without release:
          dcmanager subcloud prestage --for-sw-deploy subcloudX
      - Verify that 24.09.1 ostree commit and metadata is
        prestaged on the subcloud as the patch release has
        been deployed on the system controller
      - Check the software list command output to see the
        release states.

Story: 2010676
Task: 50659
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/924687

Change-Id: Iead7da626b31664efb153d154054ccdf17352fe3