Fixing dcmanager reporting wrong cause of failure
The dcmanager subcloud error CMD was reporting the first error message
encountered during a playbook failure, which might not reflect the
actual cause.

Updated the logic to ensure it reports the last error message,
providing a more accurate indication of the failure cause. This was
done by modifying the awk script to search for both "FATAL" and "ERROR"
messages, instead of only "FATAL".

Also updated some of the subcloud operations function to reset the error
string in the start of any operation. Most of the functions were already
doing this, but the ones added here were missing. Updated the tests
accordingly.

Test Plan:
    - PASS: Build distributedcloud package
    - PASS: Build and install a StarlingX ISO
    - PASS: Deploy a subcloud with no errors
    - PASS: By passing an invalid field in the bootstrap values for the
            subcloud, deploy it and verify that it fails. Verify that
            the "dcmanager subcloud errors" command displays the error
            correctly.

Closes-Bug: 2076196

Change-Id: I0ac4ffc0844ffb7498bef63326ee780d7acf9786
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>