Enroll retry and complete operations
This commit enables the final steps to complete a subcloud
enrollment and retry the enrollment operation in case of
a failure.

Test Plan:

PASS: Enroll a subcloud, verify that the deployment status
is set as complete.

PASS: Force an enroll failure and retry the operation,
verify that the operation finishes as complete.

Story: 2011100
Task: 50800

Change-Id: Ia3046691b50fe4106bb38512735eca275a31ea3e
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>