Update dcorch subcloud software_version after change
This commit updates the dcorch subcloud table with the new
software_version after it's modified on the dcmanager side by the
following operations:
    - software upgrade;
    - subcloud redeploy with a different version;
    - subcloud deploy install with a different version than the one
      specified during subcloud deploy create.

It also fixes an issue where the dcmanager DB subcloud software_version
was not being updated after software upgrade.

This commit also mocks the sleep duration for the 'creating vim
strategy' state to reduce the time it takes to run some of the
orchestrator tests from ~20s to ~5s.

Test Plan:
1. PASS - Run 'dcmanager subcloud deploy create' passing 22.12 as the
          release parameter and then run 'dcmanager subcloud deploy
          install' with a different version (24.09) and verify that
          the software version on dcorch changes from 22.12 to 24.09;
2. PASS - Repeat the previous test, but use 'dcmanager subcloud deploy
          resume' instead of 'install';
3. PASS - Redeploy a subcloud specifying a different subcloud version,
          verify that the dcorch DB is updated accordingly.
4. PASS - Run software upgrade orchestration and verify that both the
          dcmanager and dcorch DBs are updated accordingly.

Story: 2011106
Task: 50811

Change-Id: I104b15301107324e61d77845cf689102a993a4f3
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>