Create bulk update for audit finished
This commit creates a method to perform bulk updates for the subcloud's
end audit request to the database.
Previously, after each subcloud's audit, the database was updated to set
the audit_finished_at and audits performed. In the new implementation,
the relevant data is stored in a dictionary that is checked every 2
seconds to update the db.

Test plan:
1. PASS: Update all audits to requested in the database and verify that,
   after a successful audit, they are updated to false.
2. PASS: Restart audit services and verify that all subclouds audit are
   performed and they finish with different timestamps.
3. PASS: Set an endpoint as out-of-sync and verify that the alarms are
   cleared and reset only once for each subcloud.
4. PASS: Verify that an invalid subcloud has the audit_finished_at
   updated without being sent to the audit worker.

Story: 2011106
Task: 50878

Change-Id: I8789cf170397908177a65b70bac05e9f51de3907
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>