Fix EndpointCache KeyError
This commit addresses a potential 'KeyError' introduced by commit [1],
where accessing EndpointCache.master_service_endpoint_map[region_name]
creates an empty dictionary, due to 'master_service_endpoint_map' being
a 'defaultdict'. The code has been updated to use the '.get()' method
instead, which prevents the creation of an empty dictionary.

Additionally, this commit resolves another 'KeyError' scenario that
could occur if a new subcloud is added and the next endpoint cache
call is made when the token is close to expiring. In this situation,
the subcloud endpoint refresh would be skipped, leading to the
'KeyError'.

Test Plan:
1. PASS - Restart the dcmanager service (to clear the cached data),
create a new subcloud, and then attempt to prestage the subcloud.
Verify that the 'validate_prestage' function executes without a
'KeyError'.
2. PASS - Modify the token expiration date check to always assume that
the token will expire soon, and then re-run the previous test case.
3. PASS - Let the system run until a token is close to expiring, and
verify across all 'dcmanager' logs that there are no KeyErrors.

[1]: https://review.opendev.org/c/starlingx/distcloud/+/925132

Closes-bug: 2077480

Change-Id: I6cf11a88462027ec755b0deacd8ee3272b8db72e
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>