Handle returned error when sw-deploy-strategy is aborted
This commits handles the innacurate state that is generated
when aborting a software deploy. The operation finishes
successfully, but after running the abort command an expected
exception is raised and moves the state to "failed" when it
should remais as "aborted". A verification was added to make
sure that when this exception is raised in the "abort" flow,
its not treated as a failure.

Test Plan:

PASS: Deploy a DC system with 3 subclouds. Create a software
deploy strategy, apply it and after the strategy begins to
apply, abort it. Run the command to list the strategy-step
and verify that the subcloud with the state "aborted" does
not show an error message in details column.

Closes-Bug: 2077582

Change-Id: Ia320d0195868361c46d54914e826d0ef9921856f
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>