Fix dcorch QuotaManager EndpointCache initialization
Commit [1] introduced an issue where, if the EndpointCache is
initialized with region_name=None, it incorrectly considers it a
subcloud endpoint. This causes a failure in obtaining the subcloud
endpoints due to a missing fetch_subcloud_ips parameter.

QuotaManager was initializing the EndpointCache without a region_name,
leading to the dcorch service failure.

This commit fixes the issue by verifying that region_name is not empty
before determining if the cached master service endpoint map needs to
be refreshed. It also removes the QuotaManager 'endpoints' attribute,
as it was not being used.

[1]: https://review.opendev.org/c/starlingx/distcloud/+/926683

Test Plan:
1. PASS - Build an ISO with this commit and successfully install a
distributed cloud system;
2. PASS - Verified that there are no errors in the dcorch and dcmanager
logs.

Closes-bug: 2077651

Change-Id: I1977b28748f078b57471e91ab3e5f7bd233d3361
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>