Update software audit to handle extra releases
The software audit logic now checks if the subcloud contains more
releases in the deployed state than the SystemController. This
commit updates the audit process to correctly identify and
handle extra releases in the subcloud.

Test Plan:
PASS: Verify missing and extra releases are detected in Software
audit.
PASS: Verify that if there are no extra or missing releases, the
sync status is set to in-sync

Story: 2010676
Task: 50918

Change-Id: I82ce742b2c453d162d24f3fec1a63c605f823228
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>