Cache subcloud IDs to reduce redundant DB calls
After profiling the dcorch initial_sync process, it was seen that
the get_by_name method is consuming a large portion of the time it
takes for the process to complete. This method is needed for obtaining
the subcloud_id given a region_name, in order to create or update a
subcloud_resource.

This commit adds the subcloud_resource_id_dict dictionary to reduce
redundant get_by_name calls. The methods `get_db_subcloud_resource`
and `persist_db_subcloud_resource` were updated to use the cached
subcloud ID, improving performance by minimizing database queries.

Test Plan:
PASS: Manage a batch of subclouds for the first time on scale lab.
      Verify that the initial_sync process is completed successfully
      for all the subclouds.
PASS: Verify that the get_by_name method is called only once for each
      subcloud, and all the subsequent calls are obtaining the cached
      subcloud_id.

Story: 2011106
Task: 50945

Change-Id: Ife03b80e906a513e0420245c284633cba2c85400
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>