Update dcmanager subcloud list to consider not-available status
This commit fixes a bug introduced in [1]  where a subcloud appears
as out-of-sync even though all of its endpoints are in-sync or
not-available. This happened intermitently, because it depends on the
db_api.subcloud_get_all_with_status method to return first element of
the list as an endpoint with sync status not-available, which caused
the subcloud's sync status to be out-of-sync.

Test plan:
1. PASS: Run dcmanager subcloud list with a subcloud with all endpoints
   as in-sync and load and patching as not-available and verify that it
   returns the expected sync status of in-sync.
2. PASS: Verify that a subcloud with an endpoint out-of-sync will still
   be listed as out-of-sync when executing dcmanager subcloud list.
3. PASS: In a unit test, force the database query to return the first
   row containing an endpoint with a sync-status of not-available and
   verify that the resulting subcloud's sync status is in-sync.

[1] https://review.opendev.org/c/starlingx/distcloud/+/924454

Closes-Bug: 2078848

Change-Id: If5e7ae14d725b3f1e2d5513ec8e07d9ec6839b31
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>