Fix subcloud status after enrollment failure
This commit ensures subcloud_deploy_enroll properly returns
the success state of the operation. As a result, the 'deploy
complete' phase is now skipped when the enrollment fails.

This commit also enables deploy config phase to be ran after
enroll phase.

Test Plan:

PASS: Run subcloud enrollment. When the status is set as
'enrolling', shut down the subcloud. Verify that the playbook
fails and the status is changed to 'enroll-failed'.

PASS: Run subcloud enrollment. Verify that the subcloud is
successfully enrolled.

PASS: Run subcloud enroll with config and verify that the
phase config is executed.

Closes-bug: 2079853

Change-Id: I29095b0c10b3013820fad4f18f189e6a07272399
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>