Check the USM service availability for software deploy prestage
This commit allows us to check if the USM service is available
in a given N-1 subcloud, through which it is possible to verify the
deployed software.

The service check is performed through the Openstack driver
using the "usm" endpoint.

This change allows supporting scenarios where the subcloud has
an N-1 release with USM service available.
If the subcloud has release N-1 but the USM patch is not
installed, it will not be possible to query the USM service,
as a result the prestage will fail.

Test Plan:
Setup : - Deploy a 24.09 DC
        - Upload and deploy a 24.09.1 patch release on the system
          controller

PASS: - Deploy a 24.09 subcloud
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that 24.09.1 patch is prestaged on the subcloud

PASS: - Deploy a 22.12 subcloud with USM patch
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that release is prestaged on the subcloud

PASS: - Deploy a 22.12 subcloud without USM patch
      - Prestage the subcloud using command:
        dcmanager subcloud prestage --release 24.09
        --for-sw-deploy subcloudX
      - Verify that prestage should fail because the USM
        service is not available.

Story: 2010676
Task: 50990
Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/928976

Change-Id: Ifacdd526a202ba1ab314fa09f7af129e50eb656c
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>