Update id specification in subcloud audits bulk update
This commit updates the query introduced in [1] to bulk update the
audit_finished_at timestamp in order to fix the id being used from the
subcloud's id to the subcloud_audits' id.
Additionally, it also adds a validation in audit utils to only change
the subcloud_audits table when a value is specified, avoiding a deadlock
issue that was caused by having an empty values sent.

Test plan:
1. PASS: restart the audit process in a scale system and verify that no
   deadlock errors occur.
2. PASS: restart the system and verify that no StaleDataError occurs.
3. PASS: verify that the subcloud_audits is updated with the correct
   audit finifhed at for a subcloud.

[1] https://review.opendev.org/c/starlingx/distcloud/+/926447

Story: 2011106
Task: 50996

Change-Id: Ia69af8663cce1e99963410cfbf2efc190bffaaf1
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>